requirements.txt
pyautogen>=0.2.0
pandas
matplotlib
numpy
mistralai
python-dotenv
